requirements.txt
django-cors-headers==4.2.0
djangorestframework==3.14.0
drf-yasg==1.21.7
python-dotenv==1.0.0
requests==2.31.0
setuptools==65.5.0
paramiko==3.4.0
cassandra-driver==3.29.0
psycopg2-binary==2.9.9
minio==7.2.9
elasticsearch==8.15.1
scp==0.15.0

# sudo apt-get install sshpass